refactor: refactor: consolidate forms into reusable Form component
